fix: add css variables to root
