+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './components/iframe/iframe';
 export * from './components/material-textfield/material-textfield';
-export * from './components/nintex-styled-input/nintex-styled-input';
+export * from './components/styled-input/styled-input';
